requirements.txt
opencv-python
pillow
numpy
customtkinter
darkdetect
packaging
tkinterdnd2
